requirements.txt
requests
selenium
tensorflow
keras
pillow
numpy
pandas
matplotlib
python-dotenv